+++ cookbooks/snmp/spec/unit/resources/trapd_spec.rb
@@ -8,15 +8,23 @@ describe 'snmp_trapd' do
           platform: platform_info[:platform],
           version: platform_info[:version],
           step_into: ['snmp_trapd']
-        ).converge('snmp::snmptrapd')
+        ) do |node|
+          node.override['snmp']['community'] = 'spec-ro-community'
+          node.override['snmp']['trap']['community'] = 'spec-trap-community'
+        end.converge('snmp::snmptrapd')
       end
 
       it 'converges successfully' do
         expect { chef_run }.to_not raise_error
       end
 
-      it 'creates the SNMP trap configuration file' do
-        expect(chef_run).to create_template('/etc/snmp/snmptrapd.conf')
+      it 'writes snmptrapd.conf with mode 0600 and root ownership' do
+        expect(chef_run).to create_template('/etc/snmp/snmptrapd.conf').with(
+          mode: '0600',
+          owner: 'root',
+          group: 'root',
+          sensitive: true
+        )
       end
 
       it 'enables and starts the SNMP trap service' do
@@ -27,6 +35,18 @@ describe 'snmp_trapd' do
           expect(chef_run).to start_service('snmpd')
           expect(chef_run).to enable_service('snmpd')
         end
+      end
+
+      it 'refuses to converge with a "public" trap community' do
+        bad = ChefSpec::SoloRunner.new(
+          platform: platform_info[:platform],
+          version: platform_info[:version],
+          step_into: ['snmp_trapd']
+        ) do |node|
+          node.override['snmp']['community'] = 'spec-ro-community'
+          node.override['snmp']['trap']['community'] = 'public'
+        end
+        expect { bad.converge('snmp::snmptrapd') }.to raise_error(/public.*private/)
       end
     end
   end